refactor(utils): move selection helpers out of eval
