+++ src/sampleworks/eval/structure_utils.py
@@ -1,9 +1,8 @@
-import re
 import traceback
 from collections.abc import Sequence
 from dataclasses import dataclass, replace
 from pathlib import Path
-from typing import Any, cast, overload
+from typing import cast, overload
 
 import numpy as np
 import torch
@@ -13,12 +12,11 @@ from loguru import logger
 from sampleworks.core.rewards.protocol import RewardInputs
 from sampleworks.eval.eval_dataclasses import ProteinConfig
 from sampleworks.models.protocol import GenerativeModelInput
+from sampleworks.utils import atom_array_utils
 from sampleworks.utils.atom_array_utils import load_structure_with_altlocs
 from sampleworks.utils.atom_reconciler import AtomReconciler
 from sampleworks.utils.framework_utils import match_batch
 
-
-ATOMWORKS_COMPARISON_OPS = ("==", ">", "<", "<=", ">=", " in ")
 
 try:
     from sampleworks.models.protenix.structure_processing import (
@@ -243,114 +241,19 @@ def process_structure_to_trajectory_input(
     )
 
 
-# TODO: migrate this to atomworks's selection algebra that is added to AtomArray/Stack
-#   https://github.com/diff-use/sampleworks/issues/56
-def parse_selection_string(selection: str) -> tuple[str | None, int | None, int | None]:
-    """Parse a selection string like 'chain A and resi 326-339'.
-
-    Supports both residue ranges ('resi 10-50') and single residues ('resi 10').
-    For single residues, resi_start and resi_end will be equal.
-
-    Parameters
-    ----------
-    selection : str
-        Selection string (e.g., 'chain A', 'resi 10', 'chain A and resi 10-50')
-
-    Returns
-    -------
-    tuple
-        (chain_id, resi_start, resi_end) - any component may be None if not specified
-    """
-    # Parse "chain X and resi N-M" format and generalizations of that.
-    chain = re.search(r"chain\s+(\w+)", selection, re.IGNORECASE)
-    if chain is not None:
-        chain = chain.group(1).upper()
-
-    residues = re.search(r"resi\s+(\d+)(?:-(\d+))?", selection, re.IGNORECASE)
-    if residues is not None:
-        resi_start = int(residues.group(1))
-        resi_end = int(residues.group(2)) if residues.group(2) else resi_start
-    else:
-        resi_start = resi_end = None
-
-    if chain is None and resi_start is None and resi_end is None:
-        logger.warning(
-            "Selection string did not match any known patterns (e.g. 'chain A', 'resi 10-50')"
-        )
-
-    return chain, resi_start, resi_end
-
-
-def apply_selection(atom_array: AtomArray, selection: str | None) -> AtomArray:
-    """Apply an atom selection string to filter a structure.
-
-    Parameters
-    ----------
-    atom_array
-        Structure to filter
-    selection
-        Selection string (e.g., 'chain A and resi 10-50'). If None, returns
-        the entire structure unchanged.
-
-    Returns
-    -------
-    AtomArray
-        Filtered structure containing only atoms matching the selection
-
-    Raises
-    ------
-    ValueError
-        If the selection string matches no atoms
-    """
-    if selection is None:
-        return atom_array
-
-    if not any(x in selection for x in ATOMWORKS_COMPARISON_OPS):
-        mask = get_mask_from_old_selection_string(atom_array, selection)
-    else:
-        mask = atom_array.mask(selection)
-
-    return cast(AtomArray, atom_array[mask])
-
-
-def get_mask_from_old_selection_string(
-    atom_array: AtomArray | AtomArrayStack, selection: str
-) -> np.ndarray[tuple[int], np.dtype[Any]]:
-    DeprecationWarning(
-        f"Using old-style selection strings like {selection} is deprecated."
-        f" Use atomworks/pandas style selection strings instead."
-    )
-    chain_id, resi_start, resi_end = parse_selection_string(selection)
-    # use the length of any of the required non-coord attributes to get the mask shape
-    mask = np.ones(len(atom_array.res_id), dtype=bool)
-
-    if chain_id is not None:
-        mask &= atom_array.chain_id == chain_id
-
-    if resi_start is not None:
-        res_ids = cast(np.ndarray, atom_array.res_id)
-        if resi_end is not None:
-            mask &= (res_ids >= resi_start) & (res_ids <= resi_end)
-        else:
-            mask &= res_ids == resi_start
-
-    if mask.sum() == 0:
-        raise ValueError(f"Selection '{selection}' matched no atoms")
-    return mask
-
-
 def selection_to_residues(atom_array: AtomArray, selection: str) -> set[tuple[str, int]]:
     """Return the ``(chain_id, res_id)`` pairs covered by *selection*.
 
     Accepts both legacy ``chain X and resi a-b`` strings and atomworks-style
-    selections (``==``, ``or``, ``in``, ...) by delegating to :func:`apply_selection`,
-    so callers can treat every selection syntax uniformly. Returns an empty set
-    when the selection matches no atoms.
+    selections (``==``, ``or``, ``in``, ...) by delegating to
+    :func:`sampleworks.utils.atom_array_utils.apply_selection`, so callers can
+    treat every selection syntax uniformly. Returns an empty set when the
+    selection matches no atoms.
 
     TODO: make this work with ligands more robustly!
     """
     try:
-        selected = apply_selection(atom_array, selection)
+        selected = atom_array_utils.apply_selection(atom_array, selection)
     except ValueError:
         return set()
     if selected.array_length() == 0:
@@ -387,8 +290,8 @@ def extract_selection_coordinates(
     else:
         working_array = atom_array
 
-    if not any(x in selection for x in ATOMWORKS_COMPARISON_OPS):
-        mask = get_mask_from_old_selection_string(atom_array, selection)
+    if not any(x in selection for x in atom_array_utils.ATOMWORKS_COMPARISON_OPS):
+        mask = atom_array_utils.get_mask_from_old_selection_string(atom_array, selection)
     else:
         mask = working_array.mask(selection)
