+++ src/sampleworks/synthetic/synthetic_utils.py
@@ -9,9 +9,9 @@ from atomworks.io.transforms.atom_array import remove_waters
 from biotite.structure import AtomArray
 from loguru import logger
 
-from sampleworks.eval.structure_utils import apply_selection
 from sampleworks.utils.atom_array_utils import (
     AltlocInfo,
+    apply_selection,
     detect_altlocs,
     keep_amino_acids,
     keep_polymer,
